+++ view/list/listCategories.php
@@ -7,20 +7,24 @@
 <?php
 foreach($categories as $categorie ){ ?>
     <p><a href="index.php?ctrl=topic&action=listTopicsByCategory&id=<?= $categorie->getId() ?>"><?= $categorie->getNom() ?></a></p>
-    <p><a href="index.php?ctrl=categorie&action=modifierCategorie&id=<?=$categorie->getId()?>">Modifier</a></p>
-    <p class="myBtn">X</p>
-    <!-- The Modal -->
-    <div class="myModal modal">
-        <!-- Modal content -->
-        <div class="modal-content">
-            <p>Etes vous sur de vouloir supprimer la catégorie? Cela va entraîner la suppression de tous les topics avec leur posts!</p>
-            <a href="index.php?ctrl=categorie&action=supprimerCategorie&id=<?=$categorie->getId()?>" >Confirmer</a>
-            <div class="close">Annuler</div>
+    <?php if(App\Session::isAdmin()){ ?>
+        <p><a href="index.php?ctrl=categorie&action=modifierCategorie&id=<?=$categorie->getId()?>">Modifier</a></p>
+        <p class="myBtn">X</p>
+        <!-- The Modal -->
+        <div class="myModal modal">
+            <!-- Modal content -->
+            <div class="modal-content">
+                <p>Etes vous sur de vouloir supprimer la catégorie? Cela va entraîner la suppression de tous les topics avec leur posts!</p>
+                <a href="index.php?ctrl=categorie&action=supprimerCategorie&id=<?=$categorie->getId()?>" >Confirmer</a>
+                <div class="close">Annuler</div>
+            </div>
         </div>
-    </div>
+    <?php } ?>
+    
 <?php } ?>
 
-<a href="index.php?ctrl=categorie&action=ajouterCategorieAffichage&id="<?=$categorie->getId()?>>Nouvelle catégorie</a>
-
+<?php if(App\Session::isAdmin()){ ?>
+    <a href="index.php?ctrl=categorie&action=ajouterCategorieAffichage&id="<?=$categorie->getId()?>>Nouvelle catégorie</a>    
+<?php } ?>
